+++ database/factories/DepartmentFactory.php
@@ -3,16 +3,16 @@
 namespace Database\Factories;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
-use LaraZeus\Wind\Models\Category;
+use LaraZeus\Wind\Models\Department;
 
-class CategoryFactory extends Factory
+class DepartmentFactory extends Factory
 {
     /**
      * The name of the factory's corresponding model.
      *
      * @var string
      */
-    protected $model = Category::class;
+    protected $model = Department::class;
 
     /**
      * Define the model's default state.