+++ database/factories/LetterFactory.php
@@ -3,7 +3,7 @@
 namespace Database\Factories;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
-use LaraZeus\Wind\Models\Category;
+use LaraZeus\Wind\Models\Department;
 use LaraZeus\Wind\Models\Letter;
 
 class LetterFactory extends Factory
@@ -27,7 +27,7 @@ class LetterFactory extends Factory
         return [
             'name' => $this->faker->name,
             'email' => $this->faker->email,
-            'category_id' => Category::factory(),
+            'department_id' => Department::factory(),
             'title' => $this->faker->words(3, true),
             'message' => $this->faker->words(5, true),
             'status' => 'NEW',