+++ database/seeders/WindSeeder.php
@@ -3,19 +3,19 @@
 namespace Database\Seeders;
 
 use Illuminate\Database\Seeder;
-use LaraZeus\Wind\Models\Category;
+use LaraZeus\Wind\Models\Department;
 use LaraZeus\Wind\Models\Letter;
 
 class WindSeeder extends Seeder
 {
     public function run()
     {
-        Category::factory()->has(
+        Department::factory()->has(
             Letter::factory()
                 ->count(5)
-                ->state(function (array $attributes, Category $category) {
+                ->state(function (array $attributes, Department $department) {
                     return [
-                        'category_id' => $category->id,
+                        'department_id' => $department->id,
                     ];
                 })
         )->count(3)->create();