+++ resources/views/contact.blade.php
@@ -1,14 +1,14 @@
 <div>
     <x-slot name="title">
-        Contact us
+        {{ __('Contact us') }}
     </x-slot>
     <x-slot name="header">
-        <h2>Contact us</h2>
+        <h2>{{ __('Contact us') }}</h2>
     </x-slot>
 
     <x-zeus::box shadowless class="max-w-4xl mx-auto">
         <div class="text-gray-400 text-sm">
-            feel free to contact us.
+            {{ __('feel free to contact us.') }}
         </div>
     </x-zeus::box>
 
@@ -16,26 +16,28 @@
         @include('zeus-wind::submitted')
     @else
         <form wire:submit.prevent="store">
-            @if(config('zeus-wind.enableCategories'))
-                @if($categories->isEmpty())
-                    <input type="hidden" name="category_id" wire:model="category_id" value="1">
+            @if(config('zeus-wind.enableDepartments'))
+                @if($departments->isEmpty())
+                    <input type="hidden" name="department_id" wire:model="department_id" value="1">
                 @else
-                    @if($categories->count() > 1)
+                    @if($departments->count() > 1)
                         <div class="max-w-4xl mx-auto text-primary-600 -mb-4 mt-4">
-                            Select Category:
-                            @error('category_id') <p class="text-red-500">{{ $message }}</p> @enderror
+                            {{ __('Select Department') }}:
+                            @error('department_id') <p class="text-red-500">{{ $message }}</p> @enderror
                         </div>
                         <div class="max-w-4xl mx-auto my-6 grid grid-cols-1 sm:grid-cols-3 lg:grid-cols-4 gap-2 ">
-                            @foreach($categories as $category)
+                            @foreach($departments as $dept)
                                 <div>
                                     <label class="checkbox-wrapper">
-                                        <input wire:model="category_id" type="radio" class="checkbox-input" name="group" value="{{ $category->id }}"/>
+                                        <input wire:model="department_id" type="radio" class="checkbox-input" name="group" value="{{ $dept->id }}"/>
                                         <span class="checkbox-tile hover:border-secondary-500 p-4">
                                         <span class="text-primary-600 flex flex-col items-center justify-center gap-2">
-                                            <img class="w-full h-32 object-center object-cover" src="{{ (filter_var($category->logo, FILTER_VALIDATE_URL) === false) ? asset('storage/'.$category->logo) : $category->logo }}">
-                                            {{ $category->name ?? '' }}
+                                            @if($dept->logo !== null)
+                                                <img class="w-full h-32 object-center object-cover" src="{{ \Illuminate\Support\Facades\Storage::disk(config('zeus-wind.uploads.disk','public'))->url($dept->logo) }}">
+                                            @endif
+                                            {{ $dept->name ?? '' }}
                                         </span>
-                                        <span class="text-gray-500 text-center px-2 overflow-clip overflow-hidden ">{{ $category->desc ?? '' }}</span>
+                                        <span class="text-gray-500 text-center px-2 overflow-clip overflow-hidden ">{{ $dept->desc ?? '' }}</span>
                                     </span>
                                     </label>
                                 </div>
@@ -43,8 +45,8 @@
                         </div>
                     @else
                         <div class="max-w-4xl mx-auto text-primary-600 my-4">
-                            Sending to:
-                            <span class="text-secondary-500 font-semibold">{{ $categories->first()->name ?? '' }}</span>
+                            {{ __('Sending to') }}:
+                            <span class="text-secondary-500 font-semibold">{{ $departments->first()->name ?? '' }}</span>
                         </div>
                     @endif
                 @endif
@@ -78,7 +80,7 @@
             </x-zeus::box>
 
             <div class="p-4 text-center">
-                <x-zeus::button type="submit" class="g-recaptcha">Send</x-zeus::button>
+                <x-zeus::button type="submit" class="g-recaptcha">{{ __('Send') }}</x-zeus::button>
             </div>
 
         </form>