+++ src/Filament/Resources/DepartmentResource/Pages/CreateDepartment.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace LaraZeus\Wind\Filament\Resources\DepartmentResource\Pages;
+
+use Filament\Resources\Pages\CreateRecord;
+use LaraZeus\Wind\Filament\Resources\DepartmentResource;
+
+class CreateDepartment extends CreateRecord
+{
+    protected static string $resource = DepartmentResource::class;
+}