+++ src/Filament/Resources/DepartmentResource/Pages/EditDepartment.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace LaraZeus\Wind\Filament\Resources\DepartmentResource\Pages;
+
+use Filament\Resources\Pages\EditRecord;
+use LaraZeus\Wind\Filament\Resources\DepartmentResource;
+
+class EditDepartment extends EditRecord
+{
+    protected static string $resource = DepartmentResource::class;
+}