+++ src/Filament/Resources/DepartmentResource/Pages/ListDepartments.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace LaraZeus\Wind\Filament\Resources\DepartmentResource\Pages;
+
+use Filament\Resources\Pages\ListRecords;
+use LaraZeus\Wind\Filament\Resources\DepartmentResource;
+
+class ListDepartments extends ListRecords
+{
+    protected static string $resource = DepartmentResource::class;
+}