+++ src/Models/Department.php
@@ -2,12 +2,12 @@
 
 namespace LaraZeus\Wind\Models;
 
-use Database\Factories\CategoryFactory;
+use Database\Factories\DepartmentFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class Category extends Model
+class Department extends Model
 {
     use SoftDeletes;
     use HasFactory;
@@ -29,7 +29,7 @@ class Category extends Model
 
     protected static function newFactory()
     {
-        return CategoryFactory::new();
+        return DepartmentFactory::new();
     }
 
     public function letters()