+++ src/Models/Letter.php
@@ -13,7 +13,7 @@ class Letter extends Model
     use HasFactory;
 
     protected $fillable = [
-        'name', 'email', 'category_id', 'title', 'message', 'status', 'reply_message', 'reply_title',
+        'name', 'email', 'department_id', 'title', 'message', 'status', 'reply_message', 'reply_title',
     ];
 
     protected static function newFactory()
@@ -21,9 +21,9 @@ class Letter extends Model
         return LetterFactory::new();
     }
 
-    public function category()
+    public function department()
     {
-        return $this->belongsTo(Category::class);
+        return $this->belongsTo(Department::class);
     }
 
     public function getReplyTitleAttribute($value)